pyproject.toml
[tool.poetry]
name = "crawlall"
version = "0.1.0"
description = "Crawlall (craw-all) is a simple mail crawler tool that uses google search engine supported features to find and collect required patterns."
authors = ["ZappaBoy <federico.zappone@justanother.cloud>"]
maintainers = ["ZappaBoy <federico.zappone@justanother.cloud>"]
readme = "README.md"
packages = [{ include = "crawlall" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
googlesearch-python = "^1.2.3"
pydantic = "^2.4.2"
pandas = "^2.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
python-dotenv = "^1.0.0"
poetry2setup = "^1.1.0"
pytest-env = "^0.8.1"

[tool.pytest.ini_options]
addopts = "-s -v"
pythonpath = [
    "crawlall"
]
env = [
    "ENVIRONMENT=testing",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
test = "pytest:main"
crawlall = "crawlall:main"